pyproject.toml
[project]
name = "mastercard_payment_ops_agent_demo"
version = "0.1.0"
description = "Production-ready Agentic AI demo: Mastercard-style Payment Operations agent with LangGraph ReAct, custom MCP-like tools, Slack MCP integration, and TrueFoundry AI Gateway."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  # --- Core runtime ---
  "python-dotenv>=1.1.0",
  "pydantic>=2.8.0",
  "truefoundry>=0.10.0,<0.11.0",
  # --- API / serving ---
  "fastapi>=0.115.8",
  "uvicorn>=0.34.0",
  "starlette>=0.46.2",
  # --- Agent framework ---
  "langchain>=0.3.25",
  "langchain-openai>=0.3.9",
  "langgraph>=0.3.18",
  "langgraph-checkpoint-sqlite>=2.0.6",
  "langchain-community>=0.3.20",
  # --- Observability ---
  "traceloop-sdk==0.39.0",
  "opentelemetry-exporter-otlp>=1.36.0",
  # --- MCP + web search (optional) ---
  "mcp[cli]>=1.9.1",
  "autogen-ext[mcp,openai]>=0.4.9.2",
  "langchain-tavily>=0.2.11",
  "httpx>=0.28.1",
  # --- UI (optional) ---
  "streamlit>=1.42.1",
  "fastmcp>=3.0.1",
  "presidio-analyzer>=2.2.361",
  "presidio-anonymizer>=2.2.361",
  "pip>=26.0.1",
]

[dependency-groups]
dev = [
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
  "ipykernel>=6.29.5",
  "pre-commit>=4.1.0",
]
